* count above threshold is split into two branches * bug in Source init from cv::Mat with indexing

+++ common/misc/ThreadPool.h
@@ -86,7 +86,7 @@ namespace cmn {
 template<typename F, typename Iterator, typename Pool>
 void distribute_indexes(F&& fn, Pool& pool, Iterator start, Iterator end, uint32_t threads = 0) {
     if(threads == 0)
-        threads = pool.num_threads();
+        threads = (uint32_t)pool.num_threads();
     int64_t i = 0, N;
     if constexpr(std::integral<Iterator>)
         N = end - start;
